requirements.txt
GPUtil==1.4.0
lpips==0.1.2
matplotlib==3.7.1
numpy==1.24.3
opencv_contrib_python==4.5.3.56
opencv_python==4.5.5.62
pandas==1.4.1
Pillow==9.5.0
pytorch_msssim==0.2.1
scikit_image==0.18.3
scikit_video==1.1.11
tensorboardX==2.6
torch==1.13.1
torchvision==0.14.1
